credentials module and updated README for 3181 SDK
